feat: export getKeyManagerOptions from config
